Fix duplicated code in tests.
